Create separate component for response widget
