+++ src/components/campaign/SingleActionForm.scss
@@ -57,40 +57,8 @@
     .SingleActionForm-info {
         max-height: 10em;
     }
+
     .SingleActionForm-toggleExpandButton {
         @include circle-button($icon:$fa-var-caret-up);
     }
-}
-.SingleActionForm-response {
-    position: absolute;
-    right: 1em;
-    bottom: calc(50% - 1em);
-    color: $c-brand-main;
-    .SingleActionForm-checkbox {
-        display: none;
-        & + label {
-            display: block;
-            &:before {
-                content: ' ';
-                display: block;
-                width: 1.5em;
-                height: 1.5em;
-                background-color: darken(#fff, 1);
-                border: 1px solid darken(#fff, 15);
-                border-radius: 2px;
-                transition: background-color 0.2s;
-            }
-            span {
-                display: none;
-            }
-        }
-        &:checked +label:before {
-                background-color: darken(#fff, 5);
-                @include icon($fa-var-check);
-                height: 1.5em;
-                width: 1.5em;
-                margin: 0;
-            }
-    }
-
 }
